+++ libs/render_delegate/reader.cpp
@@ -161,12 +161,17 @@ void HydraArnoldReader::ReadStage(UsdStageRefPtr stage,
     // This creates the arnold nodes, but they don't contain any data
     SdfPathVector _excludedPrimPaths; // excluding nothing
     SdfPath rootPath = (path.empty()) ? SdfPath::AbsoluteRootPath() : SdfPath(path.c_str());
+
     UsdPrim rootPrim = stage->GetPrimAtPath(rootPath);
     _imagingDelegate->Populate(rootPrim, _excludedPrimPaths);
     if (!path.empty()) {
         UsdGeomXformCache xformCache(_imagingDelegate->GetTime());
         _imagingDelegate->SetRootTransform(xformCache.GetLocalToWorldTransform(rootPrim));
-    }    
+    }
+    // This will return a "hidden" render tag if a primitive is of a disabled type
+    _imagingDelegate->SetDisplayRender(_purpose == UsdGeomTokens->render);
+    _imagingDelegate->SetDisplayProxy(_purpose == UsdGeomTokens->proxy);
+    _imagingDelegate->SetDisplayGuides(_purpose == UsdGeomTokens->guide);
 
     // Not sure about the meaning of collection geometry -- should that be extended ?
     HdRprimCollection collection(HdTokens->geometry, HdReprSelector(HdReprTokens->hull));
@@ -209,6 +214,15 @@ void HydraArnoldReader::ReadStage(UsdStageRefPtr stage,
 
     arnoldRenderDelegate->ProcessConnections();
     
+    // We want to render the purpose that this reader was assigned to.
+    // We must also support the purpose "default". Also, when no
+    // purpose is set in the usd file, it seems to shows as "geometry", so we need to support that too
+    TfTokenVector purpose;
+    purpose.push_back(UsdGeomTokens->default_);
+    purpose.push_back(_purpose);
+    purpose.push_back(HdTokens->geometry);
+    arnoldRenderDelegate->SetRenderTags(purpose);
+
     // The scene might not be up to date, because of light links, etc, that were generated during the first sync.
     // ShouldSkipIteration updates the dirtybits for a resync, this is how it works in our hydra render pass.
     while (arnoldRenderDelegate->ShouldSkipIteration(_renderIndex, shutter)) {